mcp_server/requirements.txt
# MCP Server dependencies
mcp>=1.0.0
supabase>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
